requirements.txt
streamlit
pdfplumber
beautifulsoup4
langchain
langchain_groq
python-dotenv

